+++ src/app/components/courses-dashboard/course-details/course-tutoring-dashboard/tutoring-session-card/tutoring-session-card.component.ts
@@ -2,6 +2,8 @@ import {Component, Input, OnInit} from '@angular/core';
 import {TutoringSessionDto} from "../../../../../domain/tutoringSessionDto";
 import {TutoringSessionService} from "../../../../../services/tutoring-session.service";
 import {CourseDto} from "../../../../../domain/courseDto";
+import {ConfirmationDialogComponent} from "../../../../confirmation-dialog/confirmation-dialog.component";
+import {MatDialog} from "@angular/material/dialog";
 
 @Component({
   selector: 'tutoring-session-card',
@@ -16,7 +18,8 @@ export class TutoringSessionCardComponent implements OnInit {
   @Input() isAdmin!: boolean;
   @Input() isProfesor!: boolean;
 
-  constructor(private tutoringService: TutoringSessionService) {
+  constructor(private tutoringService: TutoringSessionService,
+              public dialog: MatDialog) {
   }
 
   ngOnInit(): void {
@@ -27,10 +30,21 @@ export class TutoringSessionCardComponent implements OnInit {
   }
 
   onDelete(tutoringSession: TutoringSessionDto) {
-    this.tutoringService.deleteById(tutoringSession.id)
-      .subscribe(() =>{
-        this.reloadPage();
-      })
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      maxWidth: "400px",
+      data: {
+        message: 'Sunteți sigur de ștergerea sesiunii de tutoring?'
+      }
+    });
+
+    dialogRef.afterClosed().subscribe(dialogResult => {
+      if (dialogResult) {
+        this.tutoringService.deleteById(tutoringSession.id)
+          .subscribe(() => {
+            this.reloadPage();
+          })
+      }
+    });
   }
 
   reloadPage() {
